move multichain constants into MultiChain folder
